Refactor daily entry tile into shared component
